test(snapshot): move assertFailClosed into test-only module, dedupe corruption cases
assertFailClosed and LEAK_PATTERNS were exported from snapshot-codec.ts,
a production module, purely for test consumption. Move them to
_internal/testing.ts alongside the other test-only exports, and update
the corruption suites' imports accordingly.

Also drop corruption-suite cases (descriptor field validation, ceiling
violations, unsupported schema/collation/log_seq_start values) that
byte-level corruption isn't actually exercising differently from the
existing coverage in snapshot-manifest.test.ts — those are ordinary
value-validation cases, not corruption cases, and were duplicated
there already.

Fixes a stale bit-position in the corrupt-UTF8-lead-byte doc comment
(0b1x0xxxxx -> 0b10xxxxxx) left over from an earlier revision.
