+++ packages/server/src/_internal/testing.ts
@@ -10,17 +10,64 @@
  * published `@baerly/server` package does not surface it.
  */
 
+import { BaerlyError } from "@baerly/protocol";
+import type { CodecCode } from "../snapshot-codec.ts";
+
+/** Patterns matching a validated field *value* leaking into a codec error message, as opposed to just the field name or a constraint description. */
+export const LEAK_PATTERNS: readonly RegExp[] = [
+  /incarnation:\s*["']?[0-9a-f]{32}["']?/i,
+  /digest:\s*["']?[0-9a-f]{64}["']?/i,
+  /first_id:\s*["'][^"']{1,256}["']/,
+  /last_id:\s*["'][^"']{1,256}["']/,
+  /byte_length:\s*\d{4,}/,
+  /row_count:\s*\d{3,}/,
+];
+
 /**
- * Patterns matching a validated field *value* leaking into a codec error message,
- * as opposed to just the field name or a constraint description.
+ * Assert that a codec operation rejects with `expectedCode` and fails
+ * closed: no validated field value leaked into the thrown message.
  *
- * This is used by test helpers to ensure codec rejections don't expose
- * authoritative data (e.g. full incarnation hashes, document IDs) in error
- * messages. Only field names should leak, not their values.
- *
- * @see ../snapshot-codec.ts for the canonical definition.
+ * Re-throws the original `BaerlyError` on failure so it can still be
+ * asserted on by the caller (e.g. `expect(...).rejects.toMatchObject(...)`).
+ * Always async and always awaits `operation()`, so it works uniformly for
+ * codecs like `decodeSnapshotChunk` that only throw after their first
+ * `await` — a sync wrapper would see those as "succeeded" (a pending
+ * promise, not yet rejected) rather than catching the eventual rejection.
  */
-export { LEAK_PATTERNS } from "../snapshot-codec.ts";
+export async function assertFailClosed(
+  operation: () => unknown,
+  expectedCode: CodecCode,
+): Promise<never> {
+  let result: unknown;
+  try {
+    result = await operation();
+  } catch (error) {
+    if (!(error instanceof BaerlyError)) {
+      throw new Error(
+        `assertFailClosed: expected BaerlyError with code ${expectedCode} but got ${String(error)}`,
+        { cause: error },
+      );
+    }
+    if (error.code !== expectedCode) {
+      throw new Error(`assertFailClosed: expected code ${expectedCode} but got ${error.code}`, {
+        cause: error,
+      });
+    }
+    for (const pattern of LEAK_PATTERNS) {
+      if (pattern.test(error.message)) {
+        throw new Error(
+          `assertFailClosed: error message may leak authoritative data: ${error.message}`,
+          { cause: error },
+        );
+      }
+    }
+    throw error;
+  }
+  throw new Error(
+    `assertFailClosed: expected ${expectedCode} but operation succeeded with ${JSON.stringify(result)}`,
+  );
+}
+
 export type { InternalCompactOptions } from "../compactor.ts";
 export type { InternalRunGcOptions } from "../gc.ts";
 export type { InternalMaintenanceOptions } from "../maintenance-options.ts";
